- Finalización de traspaso de modulo login hacia modelo vista presentador - Se configuro estilo para botones de inicio de sesion de RRSS
