+++ app/src/main/AndroidManifest.xml
@@ -12,11 +12,11 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name=".view.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
-        <activity android:name=".LoginActivity"
+        <activity android:name=".view.LoginActivity"
             android:theme="@style/FullAppTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
