+++ app/src/main/java/cl/gerardomascayano/appfiestas/view/MainActivity.java
@@ -1,4 +1,4 @@
-package cl.gerardomascayano.appfiestas;
+package cl.gerardomascayano.appfiestas.view;
 
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
@@ -8,6 +8,8 @@ import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.view.Menu;
 import android.view.MenuItem;
+
+import cl.gerardomascayano.appfiestas.R;
 
 public class MainActivity extends AppCompatActivity {
